requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
aiohttp==3.9.3
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.0
pandas==2.1.4
openai==1.12.0
requests==2.31.0
